refactor: feature公開API経由のimportに整理
feature直下のindex.tsを公開APIとして追加し、feature外からの参照を公開API経由に揃えた。
