+++ apps/web/src/features/books/index.ts
@@ -1,0 +1,1 @@
+export { BookSettings } from "./bookSettings/BookSettings"
