+++ apps/web/src/features/budgets/index.ts
@@ -1,0 +1,3 @@
+export { MonthlyBudgetUsage } from "./components/MonthlyBudgetUsage"
+export { LatestMonthlyBudget } from "./latestMonthlyBudget/LatestMonthlyBudget"
+export type { MonthlyBudgetRow } from "./types"
