+++ apps/web/src/features/categories/updateCategoryName/useUpdateCategoryName.test.tsx
@@ -1,8 +1,8 @@
 import { beforeEach, describe, expect, it, vi } from "vite-plus/test"
 
 import { act, createTestQueryClient, renderHook } from "../../../test/test-utils"
-import { paymentQueryKeys } from "../../payments/queryKeys"
-import { summaryQueryKeys } from "../../summaryByMonth/queryKeys"
+import { paymentQueryKeys } from "../../payments"
+import { summaryQueryKeys } from "../../summaryByMonth"
 import { categoryQueryKeys } from "../queryKeys"
 import { useUpdateCategoryName } from "./useUpdateCategoryName"
