+++ apps/web/src/features/payments/createPayment/CategoryField/CategoryField.tsx
@@ -8,8 +8,8 @@ import {
   CategorySelect,
   ErrorCategoryOption,
   LoadingCategoryOption,
-} from "../../../categories/components/CategorySelect"
-import { useCategories } from "../../../categories/listCategory/useCategories"
+  useCategories,
+} from "../../../categories"
 
 interface CategoryFieldProps {
   error?: boolean
