+++ apps/web/src/features/payments/deletePayment/useDeletePayment.ts
@@ -2,7 +2,7 @@ import { useMutation, useQueryClient } from "@tanstack/react-query"
 import { useCallback } from "react"
 
 import type { PaymentId } from "../../../types/payment"
-import { summaryQueryKeys } from "../../summaryByMonth/queryKeys"
+import { summaryQueryKeys } from "../../summaryByMonth"
 import { paymentQueryKeys } from "../queryKeys"
 import { removePayment } from "./removePayment"
