+++ apps/web/src/features/summaryByMonth/MonthlyTotals/MonthlyTotals.tsx
@@ -4,7 +4,7 @@ import { ErrorBoundary } from "react-error-boundary"
 
 import { toCurrency } from "../../../utils/toCurrency"
 import { useDateRange } from "../../../utils/useDateRange"
-import { MonthlyBudgetUsage } from "../../budgets/components/MonthlyBudgetUsage"
+import { MonthlyBudgetUsage } from "../../budgets"
 import { useTotalExpenditures } from "../useTotalExpenditures"
 
 function MonthlyTotals() {
